ch(css styles): improve landing page
The landing page made a bit better by making it less blank and adding
more content. Also, a login link was added to all the pages so users
could log out of the application. In addition, a login script was added
to control the flow of users on signing in - An admin will be redirected
to admin dashboard with more functionalities while non-admin users will
be redirected to dashboards with minimal functionalites.