ft(feature): login user
This route enables registered users to login to their account and
perform activities such as view their profile, vote, contest for
election.....etc. Because most of the routes are protected, a token is
generated if the user supplies a valid email and password, otherwise a
user is denied access into the application.

[Delivers #163541092]